Move orbitdb into a data directory
