+++ data/orbitdb/src/dataLoader.ts
@@ -3,8 +3,8 @@ import { readFile } from "xlsx";
 import yargs from "yargs";
 import { v4 as uuidv4 } from 'uuid';
 import { hideBin } from "yargs/helpers";
-import { EmissionsFactorInterface, EMISSIONS_FACTOR_CLASS_IDENTIFER } from "../../emissions-data/chaincode/emissionscontract/typescript/src/lib/emissionsFactor";
-import { UtilityLookupItemInterface, UTILITY_LOOKUP_ITEM_CLASS_IDENTIFIER } from "../../emissions-data/chaincode/emissionscontract/typescript/src/lib/utilityLookupItem";
+import { EmissionsFactorInterface, EMISSIONS_FACTOR_CLASS_IDENTIFER } from "../../../emissions-data/chaincode/emissionscontract/typescript/src/lib/emissionsFactor";
+import { UtilityLookupItemInterface, UTILITY_LOOKUP_ITEM_CLASS_IDENTIFIER } from "../../../emissions-data/chaincode/emissionscontract/typescript/src/lib/utilityLookupItem";
 import { COUNTRY_MAPPINGS, STATE_NAME_MAPPING } from "./abbrevToName";
 import { addCommonYargsOptions } from "./config";
 import { EMISSIONS_FACTOR_TYPE, OrbitDBService } from "./orbitDbService";
