+++ data/orbitdb/src/getData.ts
@@ -1,6 +1,6 @@
 import yargs = require('yargs')
 import { hideBin } from "yargs/helpers"
-import { EmissionsFactorInterface } from "../../emissions-data/chaincode/emissionscontract/typescript/src/lib/emissionsFactor"
+import { EmissionsFactorInterface } from "../../../emissions-data/chaincode/emissionscontract/typescript/src/lib/emissionsFactor"
 import { addCommonYargsOptions } from "./config"
 import { ActivityInterface, OrbitDBService } from "./orbitDbService"
