+++ supply-chain/.env.SAMPLE
@@ -16,7 +16,7 @@ LEDGER_ETH_NETWORK="hardhat"
 LEDGER_EMISSION_TOKEN_CONTRACT_ADDRESS="0xXXXXX"
 # needed to lookup emissions factor, must be the orbitdb address eg: zdpuAuCqi5pqTjDPXcU2LhAErAy7GSeFX37yGGqtjh1Z3mUo4
 ORBIT_DB_ADDRESS=
-# directory where the DB was previously replicated, eg: ../orbitdb/orbitdb
+# directory where the DB was previously replicated, eg: ../data/orbitdb/orbitdb
 ORBIT_DB_DIR=
 # must silence the keychain logs
 KEYCHAIN_NOLOG=Y
